+++ src/main/java/com/api/repositories/PersonaRepositoryImpl.java
@@ -12,6 +12,7 @@ import org.springframework.stereotype.Component;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 
 
 @Component
@@ -21,7 +22,7 @@ public class PersonaRepositoryImpl implements PersonaRepository {
     private DynamoDBMapper dynamoDBMapper;
 
     @Override
-    public Persona createPersona(Persona persona) {
+    public Persona createPersona(Persona persona) throws DynamoDBMappingException {
         try{
             dynamoDBMapper.save(persona);
             return persona;
@@ -31,7 +32,7 @@ public class PersonaRepositoryImpl implements PersonaRepository {
     }
 
     @Override
-    public Persona getOnePersona(String id) {
+    public Persona getOnePersona(String id) throws DynamoDBMappingException {
         try{
             Persona persona = dynamoDBMapper.load(Persona.class,id);
             return persona;
@@ -41,7 +42,7 @@ public class PersonaRepositoryImpl implements PersonaRepository {
     }
 
     @Override
-    public Persona updatePersona(Persona persona) {
+    public Persona updatePersona(Persona persona) throws DynamoDBMappingException{
         Map<String, ExpectedAttributeValue> expectedAttributeValueMap = new HashMap<>();
         ExpectedAttributeValue expectedAttributeValue = new ExpectedAttributeValue(new AttributeValue().withS(persona.getId()));
         expectedAttributeValueMap.put("id", expectedAttributeValue);
@@ -55,22 +56,42 @@ public class PersonaRepositoryImpl implements PersonaRepository {
     }
 
     @Override
-    public void deletePersona (String personaId) {
+    public void deletePersona (String id) throws DynamoDBMappingException{
         try{
-            Persona  persona = dynamoDBMapper.load(Persona.class,personaId);
-            dynamoDBMapper.delete(persona);
+            /*
+            Persona persona = dynamoDBMapper.load(Persona.class,id);;
+            if (persona != null){
+                dynamoDBMapper.delete(persona);
+            }
+            else{
+                throw new DynamoDBMappingException();
+            }
+
+             */
+
+             Optional<Persona> persona = Optional.ofNullable(getOnePersona(id));
+             if (persona.isPresent()) {
+                 dynamoDBMapper.delete(persona.get());
+             }
+             else {
+                 throw new DynamoDBMappingException();
+             }
+
         }catch (DynamoDBMappingException e){
             throw new DynamoDBMappingException(e.getMessage());
         }
+
     }
 
     @Override
-    public List<Persona> getAllPersonas() {
+    public List<Persona> getAllPersonas() throws DynamoDBMappingException{
         try{
             List<Persona> personas = dynamoDBMapper.scan(Persona.class, new DynamoDBScanExpression());
             return personas;
         }catch (DynamoDBMappingException e){
             throw new DynamoDBMappingException(e.getMessage());
         }
+
+
     }
 }
